#78 [refactor] user와 관련된 파일들을 user 패키지 내로 이동한다.
