+++ src/main/java/kusitms/backend/user/presentation/UserController.java
@@ -2,8 +2,10 @@ package kusitms.backend.user.presentation;
 
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import kusitms.backend.auth.dto.response.TokenResponseDto;
+import kusitms.backend.user.application.dto.response.TokenResponseDto;
+import kusitms.backend.user.status.AuthSuccessStatus;
 import kusitms.backend.global.dto.ApiResponse;
+import kusitms.backend.global.exception.CustomException;
 import kusitms.backend.global.util.CookieUtil;
 import kusitms.backend.user.application.UserApplicationService;
 import kusitms.backend.user.application.dto.request.CheckNicknameRequestDto;
@@ -11,7 +13,6 @@ import kusitms.backend.user.application.dto.request.SignUpRequestDto;
 import kusitms.backend.user.application.dto.response.UserInfoResponseDto;
 import kusitms.backend.user.status.UserSuccessStatus;
 import lombok.RequiredArgsConstructor;
-import org.antlr.v4.runtime.Token;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -65,6 +66,30 @@ public class UserController {
     ) {
         userApplicationService.checkNickname(request);
         return ApiResponse.onSuccess(UserSuccessStatus._OK_NOT_DUPLICATED_NICKNAME);
+    }
+
+    /**
+     * 리프레시 토큰을 통해 새로운 액세스 토큰과 리프레시 토큰을 발급받아 쿠키에 저장합니다.
+     *
+     * @param refreshToken 클라이언트로부터 전달받은 리프레시 토큰 (쿠키에서 추출)
+     * @param response     새로운 토큰을 클라이언트 쿠키에 저장하기 위한 HTTP 응답 객체
+     * @return 성공 응답
+     * @throws CustomException 리프레시 토큰이 없거나 유효하지 않을 경우 예외 발생
+     */
+    @PutMapping("/token/re-issue")
+    public ResponseEntity<ApiResponse<Void>> reIssueToken(
+            @CookieValue(value = "refreshToken", required = false) String refreshToken,
+            HttpServletResponse response
+    ) {
+
+        // AuthService를 통해 새 토큰 생성
+        TokenResponseDto tokenResponseDto = userApplicationService.reIssueToken(refreshToken);
+
+        // 새 토큰을 쿠키에 설정
+        CookieUtil.setAuthCookies(response, tokenResponseDto.accessToken(), tokenResponseDto.refreshToken(),
+                tokenResponseDto.accessTokenExpiration(), tokenResponseDto.refreshTokenExpiration());
+
+        return ApiResponse.onSuccess(AuthSuccessStatus._OK_RE_ISSUE_TOKEN);
     }
 
     /**
